Parameters, Combining rules, binary param, vol_solver, benchmark
Parameter databases for SAFTVRMie and PCSAFT have been extended
More combining rules have also been added, particularly for the epsilon parameters
Binary parameters for SAFTVRMie, PCSAFT, sPCSAFT and ogSAFT are all obtainable.
An error was identified in the vol_solver which was subsequentally ammended (normalised pressure)
For PCSAFT, benchmarking against the enthalpy of mixing was done; more benchmarking required for sPCSAFT and SAFTVRMie.